requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiomysql==0.2.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
attrs==25.3.0
certifi==2025.8.3
click==8.3.0
fastapi==0.117.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.4
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
PyMySQL==1.1.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.37.0
yarl==1.20.1
aiocron
